Change name of javascript folder to src refactor index.html.erb
