+++ app/javascript/src/PasswordHandler.js
