+++ app/javascript/src/Registration.js
