+++ app/views/layouts/application.html.erb
@@ -8,13 +8,12 @@
 
   <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
   <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
+
+
 </head>
 
 <body>
 <%= yield %>
-<%= javascript_pack_tag "Registration" %>
-<%= javascript_pack_tag "home" %>
-
 </body>
